+++ examples/MemoryBenchmark/generate_readme.py
@@ -29,7 +29,7 @@ memory and static RAM sizes were recorded. The `FEATURE_BASELINE` selection is
 the baseline, and its memory usage  numbers are subtracted from the subsequent
 `FEATURE_*` memory usage.
 
-**Version**: AceTime v2.2.3
+**Version**: AceTime v2.3-dev
 
 **DO NOT EDIT**: This file was auto-generated using `make README.md`.
 
@@ -262,16 +262,32 @@ ASCII table.
     * STM32duino Boards 2.5.0
     * ESP32 Boards 2.0.9
 
+**v2.3-dev**
+* Implement 1-second resolution in ExtendedZoneProcessor.
+    * Increases flash usage by ~1kB on 8-bit processors, but only 0-100 bytes
+      on 32-bit processors.
+* Implement CompleteZoneProcessor as a specialization of ExtendedZoneProcessor.
+    * Enables the creation of `zonedbc` database which contains all TZDB
+      timezones, for all years going back to 1844, the earliest transition in
+      the TZDB.
+    * The flash consumption of zonedbc is roughly 2X of zonedbx.
+
 # Legend
 
-* [1] Delta flash and ram consumption for `ZoneSorterByName` and
-  `ZoneSorterByOffsetAndName` are calculated by subtracting the
+* [1] Delta flash and ram consumption for `Basic ZoneSorterByName` and
+  `Basic ZoneSorterByOffsetAndName` are calculated by subtracting the
   `BasicZoneManager (1 zone)` numbers, to isolate the memory consumption
   of just the sorter classes.
-* [2] Delta flash and ram consumption for `ZoneSorterByName` and
-  `ZoneSorterByOffsetAndName` are calculated by subtracting the
+* [2] Delta flash and ram consumption for `Extended ZoneSorterByName` and
+  `Extended ZoneSorterByOffsetAndName` are calculated by subtracting the
   `ExtendedZoneManager (1 zone)` numbers, to isolate the memory
   consumption of just the sorter classes.
+* [3] Delta flash and ram consumption for `Complete ZoneSorterByName` and
+  `Complete ZoneSorterByOffsetAndName` are calculated by subtracting the
+  `CompleteZoneManager (1 zone)` numbers, to isolate the memory
+  consumption of just the sorter classes.
+* An entry of `-1` indicates that the memory usage exceeded the maximum of the
+  microcontroller and the compiler did not generate the desired information.
 
 ## Arduino Nano
 
@@ -312,9 +328,6 @@ ASCII table.
 ```
 {stm32_results}
 ```
-
-An entry of `-1` indicates that the memory usage exceeded the maximum of the
-microcontroller and the compiler did not generate the desired information.
 
 ## SAMD51 (Adafruit ItsyBitsy M4)