+++ src/ace_time/ExtendedZone.h
@@ -9,6 +9,7 @@
 #include <AceCommon.h> // KString
 #include "../zoneinfo/ZoneInfo.h"
 #include "../zoneinfo/Brokers.h"
+#include "TimeOffset.h"
 
 class Print;
 
@@ -60,10 +61,10 @@ class ExtendedZone {
     }
 
     /** Return the STDOFF of the last ZoneEra. */
-    int16_t stdOffsetMinutes() const {
+    TimeOffset stdOffset() const {
       uint8_t numEras = mZoneInfoBroker.numEras();
       extended::ZoneEraBroker zeb = mZoneInfoBroker.era(numEras - 1);
-      return zeb.offsetMinutes();
+      return TimeOffset::forSeconds(zeb.offsetSeconds());
     }
 
     /** Return the name as a KString. */